Moves `read` logic for FlexInt, FlexUInt into inherent impls
